backend/search-image/requirements.txt
openai==1.55.3
psycopg2-binary==2.9.9
